+++ domain/src/main/java/tachiyomi/domain/source/anime/model/AnimeSource.kt
@@ -1,5 +1,7 @@
 package tachiyomi.domain.source.anime.model
 
+import tachiyomi.domain.source.manga.model.SourceHealthStatus
+
 data class AnimeSource(
     val id: Long,
     val lang: String,
@@ -8,6 +10,7 @@ data class AnimeSource(
     val isStub: Boolean,
     val pin: Pins = Pins.unpinned,
     val isUsedLast: Boolean = false,
+    val healthStatus: SourceHealthStatus = SourceHealthStatus.UNKNOWN,
 ) {
 
     val visualName: String